move ActivativationsType of its own class
